Endless playlists
Implements endless playlist functionality, where the playlist loops when it has finished playing. I feel like this is a bit messy, and will likely require some refactoring if/when I decide to implement single track adds to the queue, but it works for the moment with how the playlists are set up right now.

Also removed redundant variables in player.