Metadata refactor
Refactored code to use more modular metadata structure for holding information about songs instead of just album image links. This allows for playlist updating, which is now built-in, and occurs when a user re-downloads a playlist that has already been downloaded. Still need to add in removing of songs from directory that were removed from playlist.

Changed a bunch of responses to be ephemeral in order to be more consistent.