Scheduling
Implements scheduling for the bot to be able to run 24/7. Player runs the schedule checking subroutine after each song, and moves on to the next playlist if it's the next day.

Abstracted many pieces of cog functionality into the player so that they could be easily accessed from both the player (scheduler) and cog (app command). Changed more interaction responses to be ephemeral for consistency. Added begin command. Removed unnecessary player code and superfluous import.