Close key matching
Added a check through the metadata to find keys close to the original when the original can't be found. This occurs when a song has special characters like colons or pound symbols in it that can't be saved in the file name, which causes a discrepancy between the song's filename and the actual name of the stored song in the metadata, which previously was throwing a KeyError. This should fix that.